Registration an login done

+++ RegistrationApp/Controllers/RegisterController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using RegistrationApp.Models;
 using RegistrationApp.ServiceRepository;
 using System.Diagnostics;
@@ -16,12 +17,13 @@ namespace RegistrationApp.Controllers
 
         public IActionResult Index()
         {
-            var registrations = _registrationService.GetRegistrations() ?? Enumerable.Empty<Registration>(); ;
-            return View(registrations);
+            return View();
         }
         [HttpGet]
         public IActionResult Create()
         {
+            var countries = _registrationService.GetCountries();
+            ViewData["Countries"] = countries;
             return View();
         }
         [HttpPost]
@@ -33,7 +35,7 @@ namespace RegistrationApp.Controllers
                 bool res= _registrationService.Register(registrationRequest);
                 if (res)
                 {
-                    return RedirectToAction("Index");
+                    return RedirectToAction("Login", "Account");
                 }
             }
             return View();
@@ -41,7 +43,12 @@ namespace RegistrationApp.Controllers
         [HttpGet]
         public IActionResult Edit(Guid id)
         {
-            if(id != Guid.Empty)
+            string? userIdString = HttpContext.Session.GetString("UserId");
+            if (string.IsNullOrEmpty(userIdString) || !Guid.TryParse(userIdString, out Guid userId))
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            if (id != Guid.Empty)
             {
                 Registration registrationById= _registrationService.GetRegistrationById(id);
                 if(registrationById != null)
@@ -52,29 +59,57 @@ namespace RegistrationApp.Controllers
                         Name = registrationById.Name,
                         Email = registrationById.Email,
                         Gender = registrationById.Gender,
-                        MobileNo = registrationById.MobileNo
+                        MobileNo = registrationById.MobileNo,
+                        CountryId = registrationById.CountryId,
+                        StateId = registrationById.StateId,
+                        CityId = registrationById.CityId,
+                        Countries = _registrationService.GetCountries().Select(c => new SelectListItem
+                        {
+                            Value = c.Id.ToString(),
+                            Text = c.CountryName
+                        }),
+                        States = _registrationService.GetStatesByCountry(registrationById.CountryId).Select(s => new SelectListItem
+                        {
+                            Value = s.Id.ToString(),
+                            Text = s.StateName
+                        }),
+                        Cities = _registrationService.GetCitiesByState(registrationById.StateId).Select(ci => new SelectListItem
+                        {
+                            Value = ci.Id.ToString(),
+                            Text = ci.CityName
+                        })
                     };
                     return View(updateRegistrationVM);
                 }
             }
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Login", "Account");
         }
         [HttpPost]
         public IActionResult Edit(UpdateRegistrationVM registrationRequest)
         {
-            if(registrationRequest.Id != Guid.Empty)
+            string? userIdString = HttpContext.Session.GetString("UserId");
+            if (string.IsNullOrEmpty(userIdString) || !Guid.TryParse(userIdString, out Guid userId))
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            if (registrationRequest.Id != Guid.Empty)
             {
                 bool res = _registrationService.UpdateRegistration(registrationRequest);
                 if(res)
                 {
-                    return RedirectToAction("Index");
+                    return RedirectToAction("Index", "Dashboard");
                 }
             }
-            return View();
+            return RedirectToAction("Login", "Account");
         }
         [HttpGet]
         public IActionResult Details(Guid id)
         {
+            string? userIdString = HttpContext.Session.GetString("UserId");
+            if (string.IsNullOrEmpty(userIdString) || !Guid.TryParse(userIdString, out Guid userId))
+            {
+                return RedirectToAction("Login", "Account");
+            }
             if (id != Guid.Empty)
             {
                 Registration registrationById = _registrationService.GetRegistrationById(id);
@@ -86,24 +121,57 @@ namespace RegistrationApp.Controllers
                         Name = registrationById.Name,
                         Email = registrationById.Email,
                         Gender = registrationById.Gender,
-                        MobileNo = registrationById.MobileNo
+                        MobileNo = registrationById.MobileNo,
+                        CountryId = registrationById.CountryId,
+                        StateId = registrationById.StateId,
+                        CityId = registrationById.CityId,
+                        Countries = _registrationService.GetCountries().Select(c => new SelectListItem
+                        {
+                            Value = c.Id.ToString(),
+                            Text = c.CountryName
+                        }),
+                        States = _registrationService.GetStatesByCountry(registrationById.CountryId).Select(s => new SelectListItem
+                        {
+                            Value = s.Id.ToString(),
+                            Text = s.StateName
+                        }),
+                        Cities = _registrationService.GetCitiesByState(registrationById.StateId).Select(ci => new SelectListItem
+                        {
+                            Value = ci.Id.ToString(),
+                            Text = ci.CityName
+                        })
                     };
                     return View(registrationVM);
                 }
             }
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Login", "Account");
         }
         public IActionResult Delete(Guid id)
         {
+            string? userIdString = HttpContext.Session.GetString("UserId");
+            if (string.IsNullOrEmpty(userIdString) || !Guid.TryParse(userIdString, out Guid userId))
+            {
+                return RedirectToAction("Login", "Account");
+            }
             if (id != Guid.Empty)
             {
                 bool res = _registrationService.DeleteRegistration(id);
                 if (res)
                 {
-                    return RedirectToAction("Index");
+                    return RedirectToAction("Index", "Dashboard");
                 }
             }
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Login", "Account");
+        }
+        public JsonResult GetStatesByCountry(Guid countryId)
+        {
+            var states=_registrationService.GetStatesByCountry(countryId);
+            return Json(states);
+        }
+        public JsonResult GetCitiesByState(Guid stateId)
+        {
+            var cities = _registrationService.GetCitiesByState(stateId);
+            return Json(cities);
         }
     }
 }
